3.4 review team

+++ sql/flyway/financial_operations/V1_init_DB.sql
@@ -1,7 +1,7 @@
 create table financial_operation (
                                 id bigserial not null,
-                                account_id int8 not null,
-                                operation_category_id int8,
+                                account_id bigint not null,
+                                operation_category_id bigint,
                                 isSpending boolean DEFAULT false,
                                 amount numeric(19, 2),
                                 dateTime timestamp,
